backend/pyproject.toml
[project]
name = "jobnexus"
version = "0.1.0"
description = ""
authors = [
    {name = "LoicCK",email = "loic.chauveau.kson@proton.me"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "requests (>=2.32.5,<3.0.0)",
    "google-cloud-firestore (>=2.21.0,<3.0.0)",
    "algoliasearch (<4.0)",
    "google-cloud-bigquery (>=3.40.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)",
    "ruff (>=0.14.6,<0.15.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]

[tool.black]
line-length = 88
target-version = ['py314']

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []
